i know engrish
English is a Germanic language of the Indo-European language family, originally spoken by the inhabitants of early medieval England.[3][4][5] It is named after the Angles, one of the ancient Germanic peoples that migrated from Anglia, a peninsula on the Baltic Sea (not to be confused with East Anglia), to the area of Great Britain later named after them: England. The closest living relatives of English include Scots, followed by the Low Saxon and Frisian languages. English is genealogically a West Germanic language, though its vocabulary is also hugely influenced by Old Norman French and Latin, as well as by Old Norse (a North Germanic language).[6][7][8]

English has developed over the course of more than 1,400 years. The earliest forms of English, a group of West Germanic (Ingvaeonic) dialects brought to Great Britain by Anglo-Saxon settlers in the 5th century and further mutated by Norse-speaking Viking settlers starting in the 8th and 9th centuries, are collectively called Old English. Middle English began in the late 11th century with the Norman conquest of England; this was a period in which English absorbed abundant French and Latin vocabulary through Old French: in particular, its Old Norman dialect.[9][10] Early Modern English began in the late 15th century with the introduction of the printing press to London, the printing of the King James Bible and the start of the Great Vowel Shift.[11]

Modern English has been spreading around the world since the 17th century by the worldwide influence of the British Empire and the United States. Through all types of printed and electronic media of these countries, English has become the leading language of international discourse and the lingua franca in many regions and professional contexts such as science, navigation and law.[3] Modern English grammar is the result of a gradual change from a typical Indo-European dependent-marking pattern, with a rich inflectional morphology and relatively free word order, to a mostly analytic pattern with little inflection, and a fairly fixed subject–verb–object word order.[12] Modern English relies more on auxiliary verbs and word order for the expression of complex tenses, aspect and mood, as well as passive constructions, interrogatives and some negation.

English is the most spoken language in the world (if Chinese is divided into variants)[13] and the third-most spoken native language in the world, after Standard Chinese and Spanish.[14] It is the most widely learned second language and is either the official language or one of the official languages in almost 60 sovereign states. There are more people who have learned English as a second language than there are native speakers. As of 2005, it was estimated that there were over 2 billion speakers of English.[15] English is the majority native language in the United Kingdom, the United States, Canada, Australia, New Zealand (see Anglosphere) and the Republic of Ireland, an official language and the main language of Singapore, and it is widely spoken in some areas of the Caribbean, Africa, South Asia, Southeast Asia and Oceania.[16] It is a co-official language of the United Nations, the European Union and many other world and regional international organisations. It is the most widely spoken Germanic language, accounting for at least 70% of speakers of this Indo-European branch. English speakers are called "Anglophones". There is much variability among the many accents and dialects of English used in different countries and regions in terms of phonetics and phonology, and sometimes also vocabulary, idioms, grammar, and spelling, but it does not typically prevent understanding by speakers of other dialects and accents, although mutual unintelligibility can occur at extreme ends of the dialect continuum.